Posts types refactoring

+++ src/repos/posts-repo.ts
@@ -2,15 +2,15 @@ import {LikeModelClass, PostModelClass} from "../db/db";
 import {PostDBModel} from "../models/Posts/PostModel";
 import {likesDBModel} from "../models/Comments/LikeModel";
 import {injectable} from "inversify";
+import {createObjectIdFromSting} from "../helpers/map-ObjectId";
 
 @injectable()
 export class PostsRepo {
     async deletePost(id: string): Promise<boolean> {
-        // Mongo native driver code
-        // let result = await PostModelClass.deleteOne({"id": id})
-        // return result.deletedCount === 1
 
-        const postInstance = await PostModelClass.findOne({"id": id})
+        const _id = createObjectIdFromSting(id)
+        if (_id === null) return false
+        const postInstance = await PostModelClass.findOne({"_id": _id})
         if (!postInstance) return false
 
         await postInstance.deleteOne()
@@ -24,7 +24,7 @@ export class PostsRepo {
     ): Promise<PostDBModel> {
 
         const postInstance = new PostModelClass()
-        postInstance.id = createdPost.id
+        postInstance._id = createdPost._id
         postInstance.title = createdPost.title
         postInstance.shortDescription = createdPost.shortDescription
         postInstance.content = createdPost.content
@@ -45,18 +45,10 @@ export class PostsRepo {
     }
 
     async updatePost(id: string, title: string, shortDescription: string, content: string, blogId: string): Promise<boolean> {
-        // Mongo native driver code
-        // let result = await PostModelClass.updateOne({"id": id}, {
-        //     $set: {
-        //         title: title,
-        //         shortDescription: shortDescription,
-        //         content: content,
-        //         blogId: blogId
-        //     }
-        // })
-        // return result.matchedCount === 1
 
-        const postInstance = await PostModelClass.findOne({"id": id})
+        const _id = createObjectIdFromSting(id)
+        if (_id === null) return false
+        const postInstance = await PostModelClass.findOne({"_id": _id})
         if (!postInstance) return false
 
         postInstance.title = title
